[flutter_tools] delete BuildRunnerWebCompilationProxy and WebCompilationProxy (#70914)
